t(): Refactor Google search into test methods
